+++ app/models/concerns/searchable.rb
@@ -6,8 +6,12 @@ module Searchable
 
   included do
     # Search Scope
-    def self.search(arg)
-      term = arg.to_s.downcase.gsub(/ |$/, '%')
+    def self.search(arg, match_start: true)
+      term = if match_start
+        arg.to_s.downcase.gsub(/ |$/, '%')
+      else
+        arg.to_s.downcase.gsub(/^| |$/, '%')
+      end
       terms = [term] * search_queries.count
       where search_queries.join(' or '), *terms
     end