+++ CHANGELOG.md
@@ -10,6 +10,8 @@
   - The subject report has been removed
   - Design short names are now used on mobile versions of reports
   - Overview reports for designs on multiple events can now be filtered by event
+- **Search Changes**
+  - Search results now include projects and designs that contain the search term
 - **Sheet Changes**
   - Switching the order on the sheets index no longer resets other filters
   - Sheets can now be filtered by event